+++ common/src/main/java/dk/magnusjensen/customchestmenus/commands/CommandHandler.java
@@ -27,5 +27,6 @@ public class CommandHandler {
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher) {
         ReloadMenuCommand.register(dispatcher);
         OpenMenuCommand.register(dispatcher);
+        BindingCommand.register(dispatcher);
     }
 }