+++ fabric/src/main/java/dk/magnusjensen/customchestmenus/FabricCustomChestMenus.java
@@ -18,11 +18,21 @@
 
 package dk.magnusjensen.customchestmenus;
 
+import dk.magnusjensen.customchestmenus.events.EventHandler;
 import dk.magnusjensen.customchestmenus.network.FabricNetwork;
+import dk.magnusjensen.customchestmenus.registry.FabricAttachmentRegistry;
 import dk.magnusjensen.customchestmenus.registry.FabricMenuRegistry;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
+import net.fabricmc.fabric.api.entity.event.v1.ServerEntityWorldChangeEvents;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerEntityEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
+import net.fabricmc.fabric.api.event.player.UseBlockCallback;
+import net.fabricmc.fabric.api.event.player.UseEntityCallback;
+import net.fabricmc.fabric.api.networking.v1.EntityTrackingEvents;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionResult;
 
 public class FabricCustomChestMenus implements ModInitializer {
     
@@ -37,9 +47,44 @@ public class FabricCustomChestMenus implements ModInitializer {
 
         ServerLifecycleEvents.SERVER_STARTING.register(CommonClass::loadMenus);
 
+        // Register event handlers
+        UseBlockCallback.EVENT.register((player, level, interactionHand, blockHitResult) -> {
+            var handled = EventHandler.onBlockRightClick(player, blockHitResult.getBlockPos(), interactionHand);
+            if (handled) {
+                return InteractionResult.SUCCESS;
+            }
+
+            return InteractionResult.PASS;
+        });
+
+        UseEntityCallback.EVENT.register((player, level, interactionHand, entity, entityHitResult) -> {
+            if (entityHitResult == null) {
+                return InteractionResult.PASS;
+            }
+
+            var handled = EventHandler.onEntityRightClick(player, entity, interactionHand);
+            if (handled) {
+                return InteractionResult.SUCCESS;
+            }
+
+            return InteractionResult.PASS;
+        });
+
+        ServerEntityEvents.ENTITY_UNLOAD.register((entity, serverLevel) -> EventHandler.onEntityUnload(entity));
+        EntityTrackingEvents.START_TRACKING.register((entity, serverPlayer) -> EventHandler.onPlayerStartTracking(serverPlayer, entity));
+        EntityTrackingEvents.STOP_TRACKING.register((entity, serverPlayer) -> EventHandler.onPlayerStopTracking(serverPlayer, entity));
+        ServerEntityWorldChangeEvents.AFTER_PLAYER_CHANGE_WORLD.register(((serverPlayer, origin, destination) -> {
+            EventHandler.onPlayerChangeDimension(serverPlayer, destination);
+        }));
+        ServerPlayConnectionEvents.DISCONNECT.register((handler, server) -> {
+            ServerPlayer player = handler.getPlayer();
+            EventHandler.onPlayerLeaveServer(player);
+        });
+
 
         // Static load registries
         FabricMenuRegistry.register();
+        FabricAttachmentRegistry.register();
     }