+++ neoforge/src/main/java/dk/magnusjensen/customchestmenus/registry/NeoforgeAttachmentRegistry.java
@@ -1,0 +1,42 @@
+/*
+ *     Custom Chest Menus, a Minecraft mod that allows servers to create custom chest menus.
+ *     Copyright (c) 2025  legenden (MagnusHJensen)
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package dk.magnusjensen.customchestmenus.registry;
+
+
+import dk.magnusjensen.customchestmenus.Constants;
+import dk.magnusjensen.customchestmenus.data.PlayerDataAttachment;
+import net.minecraft.resources.ResourceLocation;
+import net.neoforged.neoforge.attachment.AttachmentType;
+import net.neoforged.neoforge.registries.DeferredRegister;
+import net.neoforged.neoforge.registries.NeoForgeRegistries;
+
+import java.util.function.Supplier;
+
+public class NeoforgeAttachmentRegistry {
+    public static final DeferredRegister<AttachmentType<?>> ATTACHMENT_TYPES = DeferredRegister.create(NeoForgeRegistries.ATTACHMENT_TYPES, Constants.MOD_ID);
+
+    public static final Supplier<AttachmentType<PlayerDataAttachment>> PLAYER_DATA = ATTACHMENT_TYPES.register("player_data",
+        () -> AttachmentType.builder(() -> new PlayerDataAttachment()).build());
+
+
+    public static <T> AttachmentType<T> findById(ResourceLocation id) {
+
+        return (AttachmentType<T>) NeoForgeRegistries.ATTACHMENT_TYPES.get(id);
+    }
+}